Now available on RSS
Voilà! A push to GitHub and now this blog is available via RSS:

https://github.com/danott/allow-empty/commits.atom

Writing in vim with autowrapping at 80 characters is a different writing
experience. Reading in Reeder with the commit formatted as a code block
is a different reading experience. This is not the ideal format for
blogging. But it's interesting to embrace the constraints and try to do
something dumb on a Thursday night.