Comment moderation is built in
Every commit in GitHub gets it's own page. Here's the pages so far:

- https://github.com/danott/allow-empty/commit/283dd396bfd642d1363d94a2e736b1604467b635
- https://github.com/danott/allow-empty/commit/97be09669118e442225c47994c1bbfe19848b09a

And each page allows for comments on that commit. Comments require
authentication via GitHub. Conversations can be locked.

I was focused on RSS...but there's an entire notification system built
in. If someone wants to receive notifications about every commit on a
repository they can.

Can you comment on any file in a GitHub repository? No, you cannot.

What a fascinating idea for broadcast, opt-in subscribing, and a medium
for civil responses.