When does creation happen?
I’m continuing to muse on this experiment. It occurs to me that there is
no affordance in the GitHub interface for adding an empty commit.
Creating posts in this format MUST happen from a remote git repository.
Outside the veneers of GitHub.

Is the distinction between git and GitHub apparent to everyone? GitHub
is an interface into the raw data stored in git. With useful layers of
guiding process and organized collaboration. Usually git commits tell
the story of an artifact being produced. And GitHub pull requests tell
the story of refinements to that artifact. Both automated and human.

How would editing a post work in this format?

```
git commit --amend
git push --force
```

There would be no record that anything changed. Unless someone was
keeping a remote git repository and encountered the need to reconcile
differences.

A non-fungible log of events could ensure that all events stay on public
record. NFT? Not for this.

What is a blog? Is it the author’s words being captured and broadcast?
Does it include the commentary and interaction that get layered on top
via an interface? Does it include the presentation and interaction?
There’s probably a spectrum.

GitHub is interesting because the data store of the commits and blobs
that make up the working tree (git) is distinctly portable from the
layers of presentation, interaction, and collaboration wrapped around
those commits (GitHub). And I’m not even talking about the working tree
that makes most git repositories useful. There are layers of useful
information. In other forms of writing this editing process is usually
invisible. The remnant is a mention of thanks in an introduction or
conclusion. With git and GitHub we can see the signals of collaboration
and refinement. If one is committed to writing about their writing with
the same vigor as writing the artifact they’re trying to produce.

Using empty commits as the primary data store loses a layer of
information. It will never be clear why something changed. Which is
often a mystery in writing. And in writing about writing. And in editing
writing about writing. And what about editing writing about editing
writing?

How did an author choose to manage the tension of competing values while
editing writing about editing writing? What words did they choose to
explain how they managed the tension? Did that editing process itself
have subroutines of evaluation? Exploring divergent branches of
possibilities while considering competing values can be expressed in
git. All these layers of execution and evaluation are useful to someone
who is deeply interested in learning the crafts. It’s feedback loops all
the way down.

And then on a blog they get compressed into a single layer. Ready for
sediment to pile on or erosion to run its course.