Make the HTML representation beautiful
`git log --pretty` is pretty in the console. I want this git log as blog
to be pretty in the browser.

I dusted off my old CSS chops and made it look the way I wanted to.
There is support for light mode and dark mode.

Finding a blessed list of [solarized][] colors is more difficult than
initially imagined. I ended up in the [solarized-css][] project, looking
at a `.styl` file. I think this is `sass`?

Some regex-foo in VSCode translated the assumed-sass declarations into
CSS custom properties. I'm not using the majority of them. But I decided
to keep the translation in the source code in case I ever need to
generate them again.

I ended up reaching for ERB to make sense of my rendering. String
interpolation can get you so far. But sometimes the mental model is
worth the reach for complexity.

[solarized]: https://ethanschoonover.com/solarized/
[solarized-css]: https://github.com/thomasf/solarized-css/blob/0b4f98a4183c55e48b2827f618dbf39f54d89dd8/src/solarized-css/partials/solarized-colors.styl