Create a post from GitHub.com
Back in [09e70ee][] I said there was no affordance for authoring posts on <https://github.com>. But that's not entirely true. Using the in-browser file creation tools makes it possible. That's what I'm doing right now.

I navigated to <https://github.com/danott/allow-empty/new/main>. I'm creating an empty file in the root of the project called `.allow-commit`. I could just as easily put the number `1` in it. Or the current time. Or anything. Any change will allow me to write a commit in this tiny modal on GitHub.com. 

I assume this won't be wrapped to 80 characters. Committing the changes will let me know.

[09e70ee]: https://allow-empty.danott.website/09e70ee852f4c99d41d527880c0bb4218be02135/